moved internshipPostSkeleton component to components folder
