+++ src/components/InternshipPostSkeleton/InternshipPostSkeleton.jsx
