+++ src/components/InternshipPostSkeleton/index.js
