+++ src/pages/Internships/InternshipsList/InternshipsList.jsx
@@ -4,7 +4,7 @@ import { Card, Col, Divider, List, Row } from "antd";
 
 import Category from "components/Category";
 import InternshipPost from "components/InternshipPost";
-import InternshipPostSkeleton from "../InternshipPostSkeleton";
+import InternshipPostSkeleton from "../../../components/InternshipPostSkeleton";
 
 import { ReactComponent as IconGroup } from "../../../assets/ButtonsGroup.svg";
